SAK-12065 Functionality complete. Tidying up loose ends like layout issues etc.
git-svn-id: https://source.sakaiproject.org/svn/sam/branches/SAK-12065@38233 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo-services/src/java/org/sakaiproject/tool/assessment/facade/authz/integrated/AuthzQueriesFacade.java
@@ -298,10 +298,13 @@ public class AuthzQueriesFacade
    */
   public void removeAuthorizationByAgentQualifierAndFunction(String agentId, String qualifierId, String functionId) {
 	    String query="select a from AuthorizationData a where a.qualifierId="+qualifierId;
-	    String clause=" and a.agentId='" + agentId + "'";
+	    // String query="from AuthorizationData a where a.qualifierId="+qualifierId;
+	    String clause=" and a.agentIdString='" + agentId + "'";
 	    clause+=" and a.functionId='" + functionId + "'";
 	    List l = getHibernateTemplate().find(query+clause);
-	    getHibernateTemplate().deleteAll(l);
+	    if (l != null && l.size() > 0) {
+	    	getHibernateTemplate().deleteAll(l);
+	    }
   }
